+++ rsql-filter-integration-tests/src/test/java/com/nomendi6/rsql/it/repository/TagRepository.java
@@ -1,0 +1,9 @@
+package com.nomendi6.rsql.it.repository;
+
+import com.nomendi6.rsql.it.domain.Tag;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface TagRepository extends JpaRepository<Tag, Long> {
+}